+++ src/stack.c
@@ -1,5 +1,6 @@
 #include <stdio.h>
 #include <stdlib.h>
+#include <inttypes.h>
 
 #include "stack.h"
 
@@ -8,21 +9,21 @@ struct stack *stack_create(void)
   struct stack *stack;
 
   stack = malloc(sizeof(*stack));
-  stack->data = malloc(DEFAULT_CAPACITY * sizeof(int));
+  stack->data = malloc(DEFAULT_CAPACITY * sizeof(intptr_t));
   stack->top = -1;
   stack->capacity = DEFAULT_CAPACITY;
   stack->size = 0;
   return stack;
 }
 
-void stack_push(struct stack *stack, int value)
+void stack_push(struct stack *stack, intptr_t value)
 {
   int new_capacity;
-  int *new_data;
+  intptr_t *new_data;
 
   if (stack->top == stack->capacity - 1) {
     new_capacity = stack->capacity * 2;
-    new_data = realloc(stack->data, new_capacity * sizeof(int));
+    new_data = realloc(stack->data, new_capacity * sizeof(intptr_t));
     if (new_data == NULL)
       return;
     stack->data = new_data;
@@ -33,19 +34,19 @@ void stack_push(struct stack *stack, int value)
   stack->data[stack->top] = value;
 }
 
-int stack_pop(struct stack *stack)
+intptr_t stack_pop(struct stack *stack)
 {
-  int popped_value;
+  intptr_t popped_value;
 
   if (stack->top <= -1)
-    return -1;
+    return (intptr_t)-1;
   popped_value = stack->data[stack->top];
   stack->top--;
   stack->size--;
   return popped_value;
 }
 
-int stack_top(struct stack *stack)
+intptr_t stack_top(struct stack *stack)
 {
   return stack->data[stack->top];
 }
@@ -63,7 +64,7 @@ void stack_dump(struct stack *stack)
     printf("|       Empty       |\n");
   } else {
     for (i = 0; i < size; ++i)
-      printf("| %-7d | %-7d |\n", i, stack->data[i]);
+      printf("| %-7d | %-7" PRIdPTR " |\n", i, stack->data[i]);
   }
   printf("+---------+---------+\n");
 }
