+++ src/stack.h
@@ -4,9 +4,10 @@
 #define DEFAULT_CAPACITY 16
 
 #include <stdbool.h>
+#include <stdint.h>
 
 struct stack {
-  int *data;
+  intptr_t *data;
   int top;
   int capacity;
   int size;
@@ -23,21 +24,21 @@ struct stack *stack_create(void);
  * @param stack The stack to push the value onto.
  * @param value The value to push onto the stack.
  */
-void stack_push(struct stack *stack, int value);
+void stack_push(struct stack *stack, intptr_t value);
 
 /**
  * @brief Pops a value from the stack.
  * @param stack The stack to pop the value from.
  * @return The value popped from the stack.
  */
-int stack_pop(struct stack *stack);
+intptr_t stack_pop(struct stack *stack);
 
 /**
  * @brief Returns the top value from the stack.
  * @param stack The stack to top the value from.
  * @return The value dropped from the stack.
  */
-int stack_top(struct stack *stack);
+intptr_t stack_top(struct stack *stack);
 
 /**
  * @brief Prints the stack.
